fix: change type of amount

+++ models/walletModel.go
@@ -1,8 +1,8 @@
 package models
 
 type WalletModel struct {
-	WalletID int `gorm:"column:wallet_id,primary_key"`
-	Amount   int `gorm:"column:amount"`
+	WalletID int     `gorm:"column:wallet_id,primary_key"`
+	Amount   float64 `gorm:"column:amount"`
 }
 
 func (WalletModel) TableName() string {
